Breaking up Command file into individual class files.
