+++ InteractiveCfd_Core/Command/ButtonPress.h
@@ -1,0 +1,11 @@
+#pragma once
+#include "Command.h"
+
+class FW_API ButtonPress : public Command
+{
+    Button* m_button;
+public:
+    ButtonPress(Panel &rootPanel);
+    void Start(Button* button);
+    void End(Button* button);
+};
