+++ InteractiveCfd_Core/Command/Command.cpp
@@ -1,0 +1,19 @@
+#include "Command.h"
+#include "GraphicsManager.h"
+#include "Panel.h"
+
+Command::Command(Panel &rootPanel)
+{
+    m_rootPanel = &rootPanel;
+}
+
+Panel* Command::GetRootPanel()
+{
+    return m_rootPanel;
+}
+
+GraphicsManager* Command::GetGraphicsManager()
+{
+    return m_rootPanel->GetPanel("Graphics")->GetGraphicsManager();
+}
+
