+++ InteractiveCfd_Core/Command/Pan.h
@@ -1,0 +1,15 @@
+#pragma once
+#include "Command.h"
+
+class FW_API Pan : public Command
+{
+    float m_initialX;
+    float m_initialY;
+public:
+    Pan(Panel &rootPanel);
+    void Start(const float initialX, const float initialY);
+    void Track(const float currentX, const float currentY);
+    void End();
+};
+
+
