+++ InteractiveCfd_Core/Command/Rotate.cpp
@@ -1,0 +1,33 @@
+#include "Rotate.h"
+#include "GraphicsManager.h"
+
+Rotate::Rotate(Panel &rootPanel) : Command(rootPanel)
+{
+    m_rootPanel = &rootPanel;
+    m_state = INACTIVE;
+}
+
+void Rotate::Start(const float initialX, const float initialY)
+{
+    m_state = ACTIVE;
+    m_initialX = initialX;
+    m_initialY = initialY;
+}
+
+void Rotate::Track(const float currentX, const float currentY)
+{
+    float dx = (currentX - m_initialX)*45.f;
+    float dy = (currentY - m_initialY)*45.f;
+    if (m_state == ACTIVE)
+    {
+        GetGraphicsManager()->Rotate(dx, dy);
+    }
+    m_initialX = currentX;
+    m_initialY = currentY;
+}
+
+void Rotate::End()
+{
+    m_state = INACTIVE;
+}
+
